Added controllers, and fixed basic fillingDB

+++ src/main/java/pl/com/dariusz/giza/warehouse/repositories/BudgetsRepository.java
@@ -3,6 +3,5 @@ package pl.com.dariusz.giza.warehouse.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.com.dariusz.giza.warehouse.domain.budgets.Budget;
 
-public interface BudgetsRepository extends JpaRepository<Budget,Long> {
-
+public interface BudgetsRepository extends JpaRepository<Budget, Integer> {
 }
